Extract helpers and utils into a separate package
