+++ packages/fetchai/connections/p2p_libp2p/go.mod
@@ -4,15 +4,15 @@ go 1.13
 
 require (
 	github.com/btcsuite/btcd v0.20.1-beta
-	github.com/golang/protobuf v1.3.1
+	github.com/golang/protobuf v1.4.1
 	github.com/ipfs/go-cid v0.0.5
-	github.com/ipfs/go-datastore v0.4.4
 	github.com/joho/godotenv v1.3.0
 	github.com/libp2p/go-libp2p v0.8.3
-	github.com/libp2p/go-libp2p-autonat v0.2.2
 	github.com/libp2p/go-libp2p-circuit v0.2.2
 	github.com/libp2p/go-libp2p-core v0.5.3
+	github.com/libp2p/go-libp2p-host v0.1.0
 	github.com/libp2p/go-libp2p-kad-dht v0.7.11
+	github.com/libp2p/go-libp2p-peerstore v0.2.3
 	github.com/multiformats/go-multiaddr v0.2.1
 	github.com/multiformats/go-multihash v0.0.13
 )
